Move the logger test to the Main cli test class.
